+++ tests/integration_tests/network.py
@@ -73,9 +73,9 @@ class Geth:
 
 def setup_ethermint(path, base_port, long_timeout_commit=False):
     cfg = Path(__file__).parent / (
-        "configs/default.jsonnet"
+        "configs/long_timeout_commit.jsonnet"
         if long_timeout_commit
-        else "configs/long_timeout_commit.jsonnet"
+        else "configs/default.jsonnet"
     )
     yield from setup_custom_ethermint(path, base_port, cfg)